# Use `#except!` in `HashWithIndifferentAccess#except`